Modified for Alexa credentials in PCF
Had to figure out how to mask Alexa credentials in PCF. To do this,
we’re going to populate a EnvironmentVariables.csv file with Jenkins
and then read that file in. We’re counting on that file being there, so
I’m checking in an empty file that only has the header row populated.
Don’t check in any file that has secret keys, please!!!

I also updated Scratch code.py with code to pull from this file if it
is populated.