more flexible string comparison, new intents
Implemented a cleaner function in esdclasses.py to normalize strings
before they’re compared. This results in much greater chance of
matching after voice translation and in spite of non-standard
characters in names.

Implemented logic for query by model only, or to query by product model
in the case that either a product or model is given, and the other is
already set in the session. Also changed some “final answer” scenarios
to flask-ask statements (that should end session) rather than questions
(which leave session open)